libdonet/dcparser: Allow inheritance from multiple parent classes
I just realized that Panda's and Astron's DC parsers do not generated an
abstract syntax tree, but instead they directly manipulate the DC file
class instance that they generate on the go as the productions are
reduced. Should've noticed this fact like.. a month ago.

I feel very clueless in lots of things in life. Even at work. Is it
because I'm dumb, or acoustic? I don't know.