dcparser: Accept 1+ atomic/parameter fields in molecular fields
Also adds the Clone trait to AST structures in order to wrap the atomic
and parameter field structures in an enum so their type restriction can
be enforced and stored within the molecular field structure.

Will be getting rid of the AST soon, once I have a working 'class' for
the higher level representation of the DC File, which will be the
'class' that the Donet daemons will directly interact with to behave
accordingly to the developer's network protocol.