dcparser: Wrap DC File in Arc and Mutex for safe shared mutability
Okay. Things are getting messy. But, at least it's safe, right? I would
get rid of the 'interior' Arc<Mutex<>>'s inside DC element structs, but
even those require to be wrapped in a Mutex in order to be wrapped in
Arc's, and Arc pointers seem to be the only viable solution to keeping
multiple shared references to different DC elements within other DC
elements, which is vital for elements like molecular fields, etc.