dcparser: New CFG for explicit type casts on numeric types
Added a new type of numeric type transform, which is the explicit cast.
It is similar to an explicit cast in C, but the resulting data type
comes after the original data type, not before. (and also in C, you cant
cast a data type token because its not a variable or literal, but that
is just a DC file thing)

This should benefit client implementations that are written in
statically typed languages, such as C/C++/Rust iirc. This was not
included in the original DC language, as Disney games used Python
clients, and Python is a dynamically typed language.