dcparser: Helpful panic message for deprecated common type alias
It's common to define a type alias for the uint8 data type as 'bool'.
This is legal in Panda, but Donet provides the 'bool' data type by
default, and acts as an alias for uint8 under the hood.

This commit adds a production rule to catch this exact alias that is
common in legacy DC files, and provide a useful deprecation message to
the developer with an explanation to the parser panic.