libdonet: Use single-threaded Rc pointers instead of Arc (#11)
Parsing happens in a single thread; although the Donet daemon will not
use a single thread only, it can make clones of the DC file structure in
memory instead and focus on performance. The DC element structure also
does not get mutated after parsing, so we dont need Arc pointers which
will take a hit on performance at runtime.

Also removing traits per each DC element structure as we are not going
to be reusing those interfaces.