donet: Set global logger level from TOML configuration value
Lots of repetition due to the properties of the `DaemonLogger` struct
required to be static or const values, as it is required by the log
crate to be stored as a static.

Will look into cleaning this up some time, but this unclean code does
not at all affect performance.