fix(el): Fixed false error on extraneous data
Instead of comparing dgi.tell() and dg.size() and checking that they are
same, we instead get dgi.remaining() and make sure its equal to the
length of the buffer minus the length of the message (used buffer).

We do this because the datagram formed is always the entire buffer size,
not necessarily the message size.

Also timestamps in log entries are now on local time to match Astron's
event logger behavior.