docs: Added developer documentation under 'internal' subdir
I moved the CONTRIBUTING.md and SECURITY.md files into the Sphinx
documentation, so I removed the original files to avoid cluttering the
root directory of the repository.

I also moved the DC language specification under this directory as there
will be a more user-friendly DC file tutorial later, and this is the
formal specification, so it classifies as developer documentation.