logo: Fourth revision of Donet logo as trademark
As of this commit, I am taking ownership of the Donet™ trademark to
protect the original identity of this software project.

Please note that though I do take ownership of the trademark this does
is only related to the project's identity (brand) and not the software
itself. The software itself is free and open source for all.

If any forks of Donet are made, they cannot use the Donet trademark.