donet-core: hashgen stores hash value as i64, truncates to u32
The Astron source shows the hash value being stored as a `uint32_t`, but
clearly it "truncates" it into a `uint32_t`, hinting that it is actually
stored as a larger integer which then gets truncated.

This has to be the solution as my hash gen experiences an overflow very
fast, with just a few dozen DC elements.